🖼  wip refactoring layout
